@@ -2,7 +2,7 @@ package coolq
 
 import (
 	"encoding/hex"
-	"io/ioutil"
+	"os"
 	"path"
 	"strconv"
 	"strings"
@@ -36,15 +36,12 @@ func ToFormattedMessage(e []message.IMessageElement, groupID int64, isRaw ...boo
 func (bot *CQBot) privateMessageEvent(c *client.QQClient, m *message.PrivateMessage) {
 	bot.checkMedia(m.Elements)
 	cqm := ToStringMessage(m.Elements, 0, true)
-	if !m.Sender.IsFriend {
-		bot.oneWayMsgCache.Store(m.Sender.Uin, "")
-	}
 	id := m.Id
 	if bot.db != nil {
 		id = bot.InsertPrivateMessage(m)
 	}
 	log.Infof("收到好友 %v(%v) 的消息: %v (%v)", m.Sender.DisplayName(), m.Sender.Uin, cqm, id)
-	fm := MSG{
+	fm := global.MSG{
 		"post_type": func() string {
 			if m.Sender.Uin == bot.Client.Uin {
 				return "message_sent"
@@ -61,7 +58,7 @@ func (bot *CQBot) privateMessageEvent(c *client.QQClient, m *message.PrivateMess
 		"font":         0,
 		"self_id":      c.Uin,
 		"time":         time.Now().Unix(),
-		"sender": MSG{
+		"sender": global.MSG{
 			"user_id":  m.Sender.Uin,
 			"nickname": m.Sender.Nickname,
 			"sex":      "unknown",
@@ -69,9 +66,6 @@ func (bot *CQBot) privateMessageEvent(c *client.QQClient, m *message.PrivateMess
 		},
 	}
 	bot.dispatchEventMessage(fm)
-	if m.Sender.Uin != c.Uin {
-		c.MarkPrivateMessageReaded(m.Sender.Uin, int64(m.Time))
-	}
 }
 
 func (bot *CQBot) groupMessageEvent(c *client.QQClient, m *message.GroupMessage) {
@@ -79,12 +73,12 @@ func (bot *CQBot) groupMessageEvent(c *client.QQClient, m *message.GroupMessage)
 	for _, elem := range m.Elements {
 		if file, ok := elem.(*message.GroupFileElement); ok {
 			log.Infof("群 %v(%v) 内 %v(%v) 上传了文件: %v", m.GroupName, m.GroupCode, m.Sender.DisplayName(), m.Sender.Uin, file.Name)
-			bot.dispatchEventMessage(MSG{
+			bot.dispatchEventMessage(global.MSG{
 				"post_type":   "notice",
 				"notice_type": "group_upload",
 				"group_id":    m.GroupCode,
 				"user_id":     m.Sender.Uin,
-				"file": MSG{
+				"file": global.MSG{
 					"id":    file.Path,
 					"name":  file.Name,
 					"size":  file.Size,
@@ -109,9 +103,6 @@ func (bot *CQBot) groupMessageEvent(c *client.QQClient, m *message.GroupMessage)
 	}
 	gm["message_id"] = id
 	bot.dispatchEventMessage(gm)
-	if m.Sender.Uin != c.Uin {
-		c.MarkGroupMessageReaded(m.GroupCode, int64(m.Id))
-	}
 }
 
 func (bot *CQBot) tempMessageEvent(c *client.QQClient, e *client.TempMessageEvent) {
@@ -124,7 +115,7 @@ func (bot *CQBot) tempMessageEvent(c *client.QQClient, e *client.TempMessageEven
 		id = bot.InsertTempMessage(m.Sender.Uin, m)
 	}
 	log.Infof("收到来自群 %v(%v) 内 %v(%v) 的临时会话消息: %v", m.GroupName, m.GroupCode, m.Sender.DisplayName(), m.Sender.Uin, cqm)
-	tm := MSG{
+	tm := global.MSG{
 		"post_type":    "message",
 		"message_type": "private",
 		"sub_type":     "group",
@@ -136,7 +127,7 @@ func (bot *CQBot) tempMessageEvent(c *client.QQClient, e *client.TempMessageEven
 		"font":         0,
 		"self_id":      c.Uin,
 		"time":         time.Now().Unix(),
-		"sender": MSG{
+		"sender": global.MSG{
 			"user_id":  m.Sender.Uin,
 			"group_id": m.GroupCode,
 			"nickname": m.Sender.Nickname,
@@ -167,7 +158,7 @@ func (bot *CQBot) groupMutedEvent(c *client.QQClient, e *client.GroupMuteEvent) 
 		}
 	}
 
-	bot.dispatchEventMessage(MSG{
+	bot.dispatchEventMessage(global.MSG{
 		"post_type":   "notice",
 		"duration":    e.Time,
 		"group_id":    e.GroupCode,
@@ -190,7 +181,7 @@ func (bot *CQBot) groupRecallEvent(c *client.QQClient, e *client.GroupMessageRec
 	gid := toGlobalID(e.GroupCode, e.MessageId)
 	log.Infof("群 %v 内 %v 撤回了 %v 的消息: %v.",
 		formatGroupName(g), formatMemberName(g.FindMember(e.OperatorUin)), formatMemberName(g.FindMember(e.AuthorUin)), gid)
-	bot.dispatchEventMessage(MSG{
+	bot.dispatchEventMessage(global.MSG{
 		"post_type":   "notice",
 		"group_id":    e.GroupCode,
 		"notice_type": "group_recall",
@@ -209,7 +200,7 @@ func (bot *CQBot) groupNotifyEvent(c *client.QQClient, e client.INotifyEvent) {
 		sender := group.FindMember(notify.Sender)
 		receiver := group.FindMember(notify.Receiver)
 		log.Infof("群 %v 内 %v 戳了戳 %v", formatGroupName(group), formatMemberName(sender), formatMemberName(receiver))
-		bot.dispatchEventMessage(MSG{
+		bot.dispatchEventMessage(global.MSG{
 			"post_type":   "notice",
 			"group_id":    group.Code,
 			"notice_type": "notify",
@@ -224,7 +215,7 @@ func (bot *CQBot) groupNotifyEvent(c *client.QQClient, e client.INotifyEvent) {
 		sender := group.FindMember(notify.Sender)
 		luckyKing := group.FindMember(notify.LuckyKing)
 		log.Infof("群 %v 内 %v 的红包被抢完, %v 是运气王", formatGroupName(group), formatMemberName(sender), formatMemberName(luckyKing))
-		bot.dispatchEventMessage(MSG{
+		bot.dispatchEventMessage(global.MSG{
 			"post_type":   "notice",
 			"group_id":    group.Code,
 			"notice_type": "notify",
@@ -237,7 +228,7 @@ func (bot *CQBot) groupNotifyEvent(c *client.QQClient, e client.INotifyEvent) {
 		})
 	case *client.MemberHonorChangedNotifyEvent:
 		log.Info(notify.Content())
-		bot.dispatchEventMessage(MSG{
+		bot.dispatchEventMessage(global.MSG{
 			"post_type":   "notice",
 			"group_id":    group.Code,
 			"notice_type": "notify",
@@ -268,8 +259,12 @@ func (bot *CQBot) groupNotifyEvent(c *client.QQClient, e client.INotifyEvent) {
 func (bot *CQBot) friendNotifyEvent(c *client.QQClient, e client.INotifyEvent) {
 	friend := c.FindFriend(e.From())
 	if notify, ok := e.(*client.FriendPokeNotifyEvent); ok {
-		log.Infof("好友 %v 戳了戳你.", friend.Nickname)
-		bot.dispatchEventMessage(MSG{
+		if notify.Receiver == notify.Sender {
+			log.Infof("好友 %v 戳了戳自己.", friend.Nickname)
+		} else {
+			log.Infof("好友 %v 戳了戳你.", friend.Nickname)
+		}
+		bot.dispatchEventMessage(global.MSG{
 			"post_type":   "notice",
 			"notice_type": "notify",
 			"sub_type":    "poke",
@@ -282,6 +277,22 @@ func (bot *CQBot) friendNotifyEvent(c *client.QQClient, e client.INotifyEvent) {
 	}
 }
 
+func (bot *CQBot) memberTitleUpdatedEvent(c *client.QQClient, e *client.MemberSpecialTitleUpdatedEvent) {
+	group := c.FindGroup(e.GroupCode)
+	mem := group.FindMember(e.Uin)
+	log.Infof("群 %v(%v) 内成员 %v(%v) 获得了新的头衔: %v", group.Name, group.Code, mem.DisplayName(), mem.Uin, e.NewTitle)
+	bot.dispatchEventMessage(global.MSG{
+		"post_type":   "notice",
+		"notice_type": "notify",
+		"sub_type":    "title",
+		"group_id":    group.Code,
+		"self_id":     c.Uin,
+		"user_id":     e.Uin,
+		"time":        time.Now().Unix(),
+		"title":       e.NewTitle,
+	})
+}
+
 func (bot *CQBot) friendRecallEvent(c *client.QQClient, e *client.FriendMessageRecalledEvent) {
 	f := c.FindFriend(e.FriendUin)
 	gid := toGlobalID(e.FriendUin, e.MessageId)
@@ -290,7 +301,7 @@ func (bot *CQBot) friendRecallEvent(c *client.QQClient, e *client.FriendMessageR
 	} else {
 		log.Infof("好友 %v 撤回了消息: %v", e.FriendUin, gid)
 	}
-	bot.dispatchEventMessage(MSG{
+	bot.dispatchEventMessage(global.MSG{
 		"post_type":   "notice",
 		"notice_type": "friend_recall",
 		"self_id":     c.Uin,
@@ -306,11 +317,11 @@ func (bot *CQBot) offlineFileEvent(c *client.QQClient, e *client.OfflineFileEven
 		return
 	}
 	log.Infof("好友 %v(%v) 发送了离线文件 %v", f.Nickname, f.Uin, e.FileName)
-	bot.dispatchEventMessage(MSG{
+	bot.dispatchEventMessage(global.MSG{
 		"post_type":   "notice",
 		"notice_type": "offline_file",
 		"user_id":     e.Sender,
-		"file": MSG{
+		"file": global.MSG{
 			"name": e.FileName,
 			"size": e.FileSize,
 			"url":  e.DownloadUrl,
@@ -341,7 +352,7 @@ func (bot *CQBot) memberPermissionChangedEvent(c *client.QQClient, e *client.Mem
 		}
 		return "unset"
 	}()
-	bot.dispatchEventMessage(MSG{
+	bot.dispatchEventMessage(global.MSG{
 		"post_type":   "notice",
 		"notice_type": "group_admin",
 		"sub_type":    st,
@@ -354,7 +365,7 @@ func (bot *CQBot) memberPermissionChangedEvent(c *client.QQClient, e *client.Mem
 
 func (bot *CQBot) memberCardUpdatedEvent(c *client.QQClient, e *client.MemberCardUpdatedEvent) {
 	log.Infof("群 %v 的 %v 更新了名片 %v -> %v", formatGroupName(e.Group), formatMemberName(e.Member), e.OldCard, e.Member.CardName)
-	bot.dispatchEventMessage(MSG{
+	bot.dispatchEventMessage(global.MSG{
 		"post_type":   "notice",
 		"notice_type": "group_card",
 		"group_id":    e.Group.Code,
@@ -384,7 +395,7 @@ func (bot *CQBot) friendRequestEvent(c *client.QQClient, e *client.NewFriendRequ
 	log.Infof("收到来自 %v(%v) 的好友请求: %v", e.RequesterNick, e.RequesterUin, e.Message)
 	flag := strconv.FormatInt(e.RequestId, 10)
 	bot.friendReqCache.Store(flag, e)
-	bot.dispatchEventMessage(MSG{
+	bot.dispatchEventMessage(global.MSG{
 		"post_type":    "request",
 		"request_type": "friend",
 		"user_id":      e.RequesterUin,
@@ -398,7 +409,7 @@ func (bot *CQBot) friendRequestEvent(c *client.QQClient, e *client.NewFriendRequ
 func (bot *CQBot) friendAddedEvent(c *client.QQClient, e *client.NewFriendEvent) {
 	log.Infof("添加了新好友: %v(%v)", e.Friend.Nickname, e.Friend.Uin)
 	bot.tempSessionCache.Delete(e.Friend.Uin)
-	bot.dispatchEventMessage(MSG{
+	bot.dispatchEventMessage(global.MSG{
 		"post_type":   "notice",
 		"notice_type": "friend_add",
 		"self_id":     c.Uin,
@@ -410,7 +421,7 @@ func (bot *CQBot) friendAddedEvent(c *client.QQClient, e *client.NewFriendEvent)
 func (bot *CQBot) groupInvitedEvent(c *client.QQClient, e *client.GroupInvitedRequest) {
 	log.Infof("收到来自群 %v(%v) 内用户 %v(%v) 的加群邀请.", e.GroupName, e.GroupCode, e.InvitorNick, e.InvitorUin)
 	flag := strconv.FormatInt(e.RequestId, 10)
-	bot.dispatchEventMessage(MSG{
+	bot.dispatchEventMessage(global.MSG{
 		"post_type":    "request",
 		"request_type": "group",
 		"sub_type":     "invite",
@@ -426,7 +437,7 @@ func (bot *CQBot) groupInvitedEvent(c *client.QQClient, e *client.GroupInvitedRe
 func (bot *CQBot) groupJoinReqEvent(c *client.QQClient, e *client.UserJoinGroupRequest) {
 	log.Infof("群 %v(%v) 收到来自用户 %v(%v) 的加群请求.", e.GroupName, e.GroupCode, e.RequesterNick, e.RequesterUin)
 	flag := strconv.FormatInt(e.RequestId, 10)
-	bot.dispatchEventMessage(MSG{
+	bot.dispatchEventMessage(global.MSG{
 		"post_type":    "request",
 		"request_type": "group",
 		"sub_type":     "add",
@@ -445,11 +456,11 @@ func (bot *CQBot) otherClientStatusChangedEvent(c *client.QQClient, e *client.Ot
 	} else {
 		log.Infof("Bot 账号在客户端 %v (%v) 登出.", e.Client.DeviceName, e.Client.DeviceKind)
 	}
-	bot.dispatchEventMessage(MSG{
+	bot.dispatchEventMessage(global.MSG{
 		"post_type":   "notice",
 		"notice_type": "client_status",
 		"online":      e.Online,
-		"client": MSG{
+		"client": global.MSG{
 			"app_id":      e.Client.AppId,
 			"device_name": e.Client.DeviceName,
 			"device_kind": e.Client.DeviceKind,
@@ -482,7 +493,7 @@ func (bot *CQBot) groupEssenceMsg(c *client.QQClient, e *client.GroupDigestEvent
 	if e.OperatorUin == bot.Client.Uin {
 		return
 	}
-	bot.dispatchEventMessage(MSG{
+	bot.dispatchEventMessage(global.MSG{
 		"post_type":   "notice",
 		"group_id":    e.GroupCode,
 		"notice_type": "essence",
@@ -500,8 +511,8 @@ func (bot *CQBot) groupEssenceMsg(c *client.QQClient, e *client.GroupDigestEvent
 	})
 }
 
-func (bot *CQBot) groupIncrease(groupCode, operatorUin, userUin int64) MSG {
-	return MSG{
+func (bot *CQBot) groupIncrease(groupCode, operatorUin, userUin int64) global.MSG {
+	return global.MSG{
 		"post_type":   "notice",
 		"notice_type": "group_increase",
 		"group_id":    groupCode,
@@ -513,8 +524,8 @@ func (bot *CQBot) groupIncrease(groupCode, operatorUin, userUin int64) MSG {
 	}
 }
 
-func (bot *CQBot) groupDecrease(groupCode, userUin int64, operator *client.GroupMemberInfo) MSG {
-	return MSG{
+func (bot *CQBot) groupDecrease(groupCode, userUin int64, operator *client.GroupMemberInfo) global.MSG {
+	return global.MSG{
 		"post_type":   "notice",
 		"notice_type": "group_decrease",
 		"group_id":    groupCode,
@@ -542,59 +553,26 @@ func (bot *CQBot) groupDecrease(groupCode, userUin int64, operator *client.Group
 func (bot *CQBot) checkMedia(e []message.IMessageElement) {
 	for _, elem := range e {
 		switch i := elem.(type) {
-		case *message.ImageElement:
-			filename := hex.EncodeToString(i.Md5) + ".image"
-			if !global.PathExists(path.Join(global.ImagePath, filename)) {
-				_ = ioutil.WriteFile(path.Join(global.ImagePath, filename), binary.NewWriterF(func(w *binary.Writer) {
-					w.Write(i.Md5)
-					w.WriteUInt32(uint32(i.Size))
-					w.WriteString(i.Filename)
-					w.WriteString(i.Url)
-				}), 0o644)
-			}
-			i.Filename = filename
 		case *message.GroupImageElement:
 			filename := hex.EncodeToString(i.Md5) + ".image"
 			if !global.PathExists(path.Join(global.ImagePath, filename)) {
-				_ = ioutil.WriteFile(path.Join(global.ImagePath, filename), binary.NewWriterF(func(w *binary.Writer) {
+				_ = os.WriteFile(path.Join(global.ImagePath, filename), binary.NewWriterF(func(w *binary.Writer) {
 					w.Write(i.Md5)
 					w.WriteUInt32(uint32(i.Size))
-					w.WriteString(filename)
+					w.WriteString(i.ImageId)
 					w.WriteString(i.Url)
 				}), 0o644)
 			}
 		case *message.FriendImageElement:
 			filename := hex.EncodeToString(i.Md5) + ".image"
 			if !global.PathExists(path.Join(global.ImagePath, filename)) {
-				_ = ioutil.WriteFile(path.Join(global.ImagePath, filename), binary.NewWriterF(func(w *binary.Writer) {
+				_ = os.WriteFile(path.Join(global.ImagePath, filename), binary.NewWriterF(func(w *binary.Writer) {
 					w.Write(i.Md5)
-					w.WriteUInt32(uint32(0)) // 发送时会调用url, 大概没事
-					w.WriteString(filename)
+					w.WriteUInt32(uint32(i.Size))
+					w.WriteString(i.ImageId)
 					w.WriteString(i.Url)
 				}), 0o644)
 			}
-		case *message.GroupFlashImgElement:
-			filename := hex.EncodeToString(i.Md5) + ".image"
-			if !global.PathExists(path.Join(global.ImagePath, filename)) {
-				_ = ioutil.WriteFile(path.Join(global.ImagePath, filename), binary.NewWriterF(func(w *binary.Writer) {
-					w.Write(i.Md5)
-					w.WriteUInt32(uint32(i.Size))
-					w.WriteString(i.Filename)
-					w.WriteString("")
-				}), 0o644)
-			}
-			i.Filename = filename
-		case *message.FriendFlashImgElement:
-			filename := hex.EncodeToString(i.Md5) + ".image"
-			if !global.PathExists(path.Join(global.ImagePath, filename)) {
-				_ = ioutil.WriteFile(path.Join(global.ImagePath, filename), binary.NewWriterF(func(w *binary.Writer) {
-					w.Write(i.Md5)
-					w.WriteUInt32(uint32(i.Size))
-					w.WriteString(i.Filename)
-					w.WriteString("")
-				}), 0o644)
-			}
-			i.Filename = filename
 		case *message.VoiceElement:
 			i.Name = strings.ReplaceAll(i.Name, "{", "")
 			i.Name = strings.ReplaceAll(i.Name, "}", "")
@@ -604,12 +582,12 @@ func (bot *CQBot) checkMedia(e []message.IMessageElement) {
 					log.Warnf("语音文件 %v 下载失败: %v", i.Name, err)
 					continue
 				}
-				_ = ioutil.WriteFile(path.Join(global.VoicePath, i.Name), b, 0o644)
+				_ = os.WriteFile(path.Join(global.VoicePath, i.Name), b, 0o644)
 			}
 		case *message.ShortVideoElement:
 			filename := hex.EncodeToString(i.Md5) + ".video"
 			if !global.PathExists(path.Join(global.VideoPath, filename)) {
-				_ = ioutil.WriteFile(path.Join(global.VideoPath, filename), binary.NewWriterF(func(w *binary.Writer) {
+				_ = os.WriteFile(path.Join(global.VideoPath, filename), binary.NewWriterF(func(w *binary.Writer) {
 					w.Write(i.Md5)
 					w.Write(i.ThumbMd5)
 					w.WriteUInt32(uint32(i.Size))
