@@ -2,8 +2,8 @@ package global
 
 import (
 	"crypto/md5"
-	"fmt"
-	"io/ioutil"
+	"encoding/hex"
+	"os"
 	"os/exec"
 	"path"
 
@@ -19,9 +19,9 @@ func EncoderSilk(data []byte) ([]byte, error) {
 	if err != nil {
 		return nil, errors.Wrap(err, "calc md5 failed")
 	}
-	tempName := fmt.Sprintf("%x", h.Sum(nil))
+	tempName := hex.EncodeToString(h.Sum(nil))
 	if silkPath := path.Join("data/cache", tempName+".silk"); PathExists(silkPath) {
-		return ioutil.ReadFile(silkPath)
+		return os.ReadFile(silkPath)
 	}
 	slk, err := codec.EncodeToSilk(data, tempName, true)
 	if err != nil {
