@@ -1,11 +1,12 @@
-// +build linux windows,!arm darwin
+//go:build (linux || (windows && !arm && !arm64) || darwin) && (386 || amd64 || arm || arm64) && !race && !nosilk
+// +build linux windows,!arm,!arm64 darwin
 // +build 386 amd64 arm arm64
+// +build !race
+// +build !nosilk
 
-// Package codec Slik编码核心模块
 package codec
 
 import (
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"path"
@@ -14,15 +15,13 @@ import (
 	"github.com/wdvxdr1123/go-silk"
 )
 
-const (
-	silkCachePath = "data/cache"
-)
+const silkCachePath = "data/cache"
 
 // EncodeToSilk 将音频编码为Silk
 func EncodeToSilk(record []byte, tempName string, useCache bool) (silkWav []byte, err error) {
 	// 1. 写入缓存文件
 	rawPath := path.Join(silkCachePath, tempName+".wav")
-	err = ioutil.WriteFile(rawPath, record, os.ModePerm)
+	err = os.WriteFile(rawPath, record, os.ModePerm)
 	if err != nil {
 		return nil, errors.Wrap(err, "write temp file error")
 	}
@@ -31,13 +30,17 @@ func EncodeToSilk(record []byte, tempName string, useCache bool) (silkWav []byte
 	// 2.转换pcm
 	pcmPath := path.Join(silkCachePath, tempName+".pcm")
 	cmd := exec.Command("ffmpeg", "-i", rawPath, "-f", "s16le", "-ar", "24000", "-ac", "1", pcmPath)
+	if Debug {
+		cmd.Stdout = os.Stdout
+		cmd.Stderr = os.Stderr
+	}
 	if err = cmd.Run(); err != nil {
 		return nil, errors.Wrap(err, "convert pcm file error")
 	}
 	defer os.Remove(pcmPath)
 
 	// 3. 转silk
-	pcm, err := ioutil.ReadFile(pcmPath)
+	pcm, err := os.ReadFile(pcmPath)
 	if err != nil {
 		return nil, errors.Wrap(err, "read pcm file err")
 	}
@@ -47,7 +50,7 @@ func EncodeToSilk(record []byte, tempName string, useCache bool) (silkWav []byte
 	}
 	if useCache {
 		silkPath := path.Join(silkCachePath, tempName+".silk")
-		err = ioutil.WriteFile(silkPath, silkWav, 0o666)
+		err = os.WriteFile(silkPath, silkWav, 0o666)
 	}
 	return
 }
