@@ -3,7 +3,6 @@ package main
 import (
 	"bufio"
 	"bytes"
-	"io/ioutil"
 	"os"
 	"strings"
 	"time"
@@ -63,9 +62,13 @@ func qrcodeLogin() error {
 	if err != nil {
 		return err
 	}
-	_ = ioutil.WriteFile("qrcode.png", rsp.ImageData, 0o644)
+	_ = os.WriteFile("qrcode.png", rsp.ImageData, 0o644)
 	defer func() { _ = os.Remove("qrcode.png") }()
-	log.Infof("请使用手机QQ扫描二维码 (qrcode.png) : ")
+	if cli.Uin != 0 {
+		log.Infof("请使用账号 %v 登录手机QQ扫描二维码 (qrcode.png) : ", cli.Uin)
+	} else {
+		log.Infof("请使用手机QQ扫描二维码 (qrcode.png) : ")
+	}
 	time.Sleep(time.Second)
 	qrcodeTerminal.New().Get(fi.Content).Print()
 	s, err := cli.QueryQRCodeStatus(rsp.Sig)
@@ -114,27 +117,14 @@ func loginResponseProcessor(res *client.LoginResponse) error {
 		var text string
 		switch res.Error {
 		case client.SliderNeededError:
-			log.Warnf("登录需要滑条验证码. ")
-			log.Warnf("请参考文档 -> https://docs.go-cqhttp.org/faq/slider.html <- 进行处理")
-			log.Warnf("1. 自行抓包并获取 Ticket 输入.")
-			log.Warnf("2. 使用手机QQ扫描二维码登入. (推荐)")
-			log.Warn("请输入(1 - 2) (将在10秒后自动选择2)：")
-			text = readLineTimeout(time.Second*10, "2")
-			if strings.Contains(text, "1") {
-				println()
-				log.Warnf("请用浏览器打开 -> %v <- 并获取Ticket.", res.VerifyUrl)
-				println()
-				log.Warn("请输入Ticket： (Enter 提交)")
-				text = readLine()
-				res, err = cli.SubmitTicket(text)
-				continue
-			}
+			log.Warnf("登录需要滑条验证码, 请使用手机QQ扫描二维码以继续登录.")
 			cli.Disconnect()
+			cli.Release()
 			cli = client.NewClientEmpty()
 			return qrcodeLogin()
 		case client.NeedCaptcha:
 			log.Warnf("登录需要验证码.")
-			_ = ioutil.WriteFile("captcha.jpg", res.CaptchaImage, 0o644)
+			_ = os.WriteFile("captcha.jpg", res.CaptchaImage, 0o644)
 			log.Warnf("请输入验证码 (captcha.jpg)： (Enter 提交)")
 			text = readLine()
 			global.DelFile("captcha.jpg")
