@@ -11,7 +11,7 @@ var bufferPool = sync.Pool{
 	},
 }
 
-// NewBuffer 从池钟获取新 bytes.Buffer
+// NewBuffer 从池中获取新 bytes.Buffer
 func NewBuffer() *bytes.Buffer {
 	return bufferPool.Get().(*bytes.Buffer)
 }
@@ -20,7 +20,7 @@ func NewBuffer() *bytes.Buffer {
 func PutBuffer(buf *bytes.Buffer) {
 	// See https://golang.org/issue/23199
 	const maxSize = 1 << 16
-	if buf.Cap() < maxSize { // 对于大Buffer直接丢弃
+	if buf != nil && buf.Cap() < maxSize { // 对于大Buffer直接丢弃
 		buf.Reset()
 		bufferPool.Put(buf)
 	}
