@@ -7,8 +7,11 @@ import (
 	"fmt"
 	"os"
 	"path"
+	"strconv"
 	"strings"
 	"sync"
+
+	"github.com/Mrs4s/go-cqhttp/global"
 
 	log "github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v3"
@@ -53,11 +56,14 @@ type Config struct {
 		ReportSelfMessage   bool   `yaml:"report-self-message"`
 		RemoveReplyAt       bool   `yaml:"remove-reply-at"`
 		ExtraReplyData      bool   `yaml:"extra-reply-data"`
+		SkipMimeScan        bool   `yaml:"skip-mime-scan"`
 	} `yaml:"message"`
 
 	Output struct {
-		LogLevel string `yaml:"log-level"`
-		Debug    bool   `yaml:"debug"`
+		LogLevel    string `yaml:"log-level"`
+		LogAging    int    `yaml:"log-aging"`
+		LogForceNew bool   `yaml:"log-force-new"`
+		Debug       bool   `yaml:"debug"`
 	} `yaml:"output"`
 
 	Servers  []map[string]yaml.Node `yaml:"servers"`
@@ -77,11 +83,15 @@ type MiddleWares struct {
 
 // HTTPServer HTTP通信相关配置
 type HTTPServer struct {
-	Disabled bool   `yaml:"disabled"`
-	Host     string `yaml:"host"`
-	Port     int    `yaml:"port"`
-	Timeout  int32  `yaml:"timeout"`
-	Post     []struct {
+	Disabled    bool   `yaml:"disabled"`
+	Host        string `yaml:"host"`
+	Port        int    `yaml:"port"`
+	Timeout     int32  `yaml:"timeout"`
+	LongPolling struct {
+		Enabled      bool `yaml:"enabled"`
+		MaxQueueSize int  `yaml:"max-queue-size"`
+	} `yaml:"long-polling"`
+	Post []struct {
 		URL    string `yaml:"url"`
 		Secret string `yaml:"secret"`
 	}
@@ -116,6 +126,14 @@ type WebsocketReverse struct {
 	MiddleWares `yaml:"middlewares"`
 }
 
+// LambdaServer 云函数配置
+type LambdaServer struct {
+	Disabled bool   `yaml:"disabled"`
+	Type     string `yaml:"type"`
+
+	MiddleWares `yaml:"middlewares"`
+}
+
 // LevelDBConfig leveldb 相关配置
 type LevelDBConfig struct {
 	Enable bool `yaml:"enable"`
@@ -129,15 +147,93 @@ var (
 // Get 从默认配置文件路径中获取
 func Get() *Config {
 	once.Do(func() {
+		hasEnvironmentConf := os.Getenv("GCQ_UIN") != ""
+
 		file, err := os.Open(DefaultConfigFile)
-		if err != nil {
+		config = &Config{}
+		if err == nil {
+			defer func() { _ = file.Close() }()
+			if err = yaml.NewDecoder(file).Decode(config); err != nil && !hasEnvironmentConf {
+				log.Fatal("配置文件不合法!", err)
+			}
+		} else if !hasEnvironmentConf {
 			generateConfig()
 			os.Exit(0)
 		}
-		defer func() { _ = file.Close() }()
-		config = &Config{}
-		if err = yaml.NewDecoder(file).Decode(config); err != nil {
-			log.Fatal("配置文件不合法!", err)
+		if hasEnvironmentConf {
+			// type convert tools
+			toInt64 := func(str string) int64 {
+				i, _ := strconv.ParseInt(str, 10, 64)
+				return i
+			}
+
+			// load config from environment variable
+			global.SetAtDefault(&config.Account.Uin, toInt64(os.Getenv("GCQ_UIN")), int64(0))
+			global.SetAtDefault(&config.Account.Password, os.Getenv("GCQ_PWD"), "")
+			global.SetAtDefault(&config.Account.Status, int32(toInt64(os.Getenv("GCQ_STATUS"))), int32(0))
+			global.SetAtDefault(&config.Account.ReLogin.Disabled, !global.EnsureBool(os.Getenv("GCQ_RELOGIN_DISABLED"), true), false)
+			global.SetAtDefault(&config.Account.ReLogin.Delay, uint(toInt64(os.Getenv("GCQ_RELOGIN_DELAY"))), uint(0))
+			global.SetAtDefault(&config.Account.ReLogin.MaxTimes, uint(toInt64(os.Getenv("GCQ_RELOGIN_MAX_TIMES"))), uint(0))
+			dbConf := &LevelDBConfig{Enable: global.EnsureBool(os.Getenv("GCQ_LEVELDB"), true)}
+			if config.Database == nil {
+				config.Database = make(map[string]yaml.Node)
+			}
+			config.Database["leveldb"] = func() yaml.Node {
+				n := &yaml.Node{}
+				_ = n.Encode(dbConf)
+				return *n
+			}()
+			accessTokenEnv := os.Getenv("GCQ_ACCESS_TOKEN")
+			if os.Getenv("GCQ_HTTP_PORT") != "" {
+				node := &yaml.Node{}
+				httpConf := &HTTPServer{
+					Host: "0.0.0.0",
+					Port: 5700,
+					MiddleWares: MiddleWares{
+						AccessToken: accessTokenEnv,
+					},
+				}
+				global.SetExcludeDefault(&httpConf.Disabled, global.EnsureBool(os.Getenv("GCQ_HTTP_DISABLE"), false), false)
+				global.SetExcludeDefault(&httpConf.Host, os.Getenv("GCQ_HTTP_HOST"), "")
+				global.SetExcludeDefault(&httpConf.Port, int(toInt64(os.Getenv("GCQ_HTTP_PORT"))), 0)
+				if os.Getenv("GCQ_HTTP_POST_URL") != "" {
+					httpConf.Post = append(httpConf.Post, struct {
+						URL    string `yaml:"url"`
+						Secret string `yaml:"secret"`
+					}{os.Getenv("GCQ_HTTP_POST_URL"), os.Getenv("GCQ_HTTP_POST_SECRET")})
+				}
+				_ = node.Encode(httpConf)
+				config.Servers = append(config.Servers, map[string]yaml.Node{"http": *node})
+			}
+			if os.Getenv("GCQ_WS_PORT") != "" {
+				node := &yaml.Node{}
+				wsServerConf := &WebsocketServer{
+					Host: "0.0.0.0",
+					Port: 6700,
+					MiddleWares: MiddleWares{
+						AccessToken: accessTokenEnv,
+					},
+				}
+				global.SetExcludeDefault(&wsServerConf.Disabled, global.EnsureBool(os.Getenv("GCQ_WS_DISABLE"), false), false)
+				global.SetExcludeDefault(&wsServerConf.Host, os.Getenv("GCQ_WS_HOST"), "")
+				global.SetExcludeDefault(&wsServerConf.Port, int(toInt64(os.Getenv("GCQ_WS_PORT"))), 0)
+				_ = node.Encode(wsServerConf)
+				config.Servers = append(config.Servers, map[string]yaml.Node{"ws": *node})
+			}
+			if os.Getenv("GCQ_RWS_API") != "" || os.Getenv("GCQ_RWS_EVENT") != "" || os.Getenv("GCQ_RWS_UNIVERSAL") != "" {
+				node := &yaml.Node{}
+				rwsConf := &WebsocketReverse{
+					MiddleWares: MiddleWares{
+						AccessToken: accessTokenEnv,
+					},
+				}
+				global.SetExcludeDefault(&rwsConf.Disabled, global.EnsureBool(os.Getenv("GCQ_RWS_DISABLE"), false), false)
+				global.SetExcludeDefault(&rwsConf.API, os.Getenv("GCQ_RWS_API"), "")
+				global.SetExcludeDefault(&rwsConf.Event, os.Getenv("GCQ_RWS_EVENT"), "")
+				global.SetExcludeDefault(&rwsConf.Universal, os.Getenv("GCQ_RWS_UNIVERSAL"), "")
+				_ = node.Encode(rwsConf)
+				config.Servers = append(config.Servers, map[string]yaml.Node{"ws-reverse": *node})
+			}
 		}
 	})
 	return config
@@ -162,6 +258,7 @@ func generateConfig() {
 > 2: 正向 Websocket 通信
 > 3: 反向 Websocket 通信
 > 4: pprof 性能分析服务器
+> 5: 云函数服务
 请输入你需要的编号，可输入多个，同一编号也可输入多个(如: 233)
 您的选择是:`)
 	input := bufio.NewReader(os.Stdin)
@@ -179,6 +276,8 @@ func generateConfig() {
 			sb.WriteString(wsReverseDefault)
 		case '4':
 			sb.WriteString(pprofDefault)
+		case '5':
+			sb.WriteString(lambdaDefault)
 		}
 	}
 	_ = os.WriteFile("config.yml", []byte(sb.String()), 0o644)
@@ -195,6 +294,12 @@ const httpDefault = `  # HTTP 通信设置
       # 反向HTTP超时时间, 单位秒
       # 最小值为5，小于5将会忽略本项设置
       timeout: 5
+      # 长轮询拓展
+      long-polling:
+        # 是否开启
+        enabled: false
+        # 消息队列大小，0 表示不限制队列大小，谨慎使用
+        max-queue-size: 2000
       middlewares:
         <<: *default # 引用默认中间件
       # 反向HTTP POST地址列表
@@ -203,6 +308,13 @@ const httpDefault = `  # HTTP 通信设置
       #  secret: ''           # 密钥
       #- url: 127.0.0.1:5701 # 地址
       #  secret: ''          # 密钥
+`
+
+const lambdaDefault = `  # LambdaServer 配置
+  - lambda:
+      type: scf # scf: 腾讯云函数 aws: aws Lambda
+      middlewares:
+        <<: *default # 引用默认中间件
 `
 
 const wsDefault = `  # 正向WS设置
