@@ -1,33 +1,33 @@
 package server
 
 import (
-	"bytes"
 	"context"
 	"crypto/hmac"
 	"crypto/sha1"
 	"encoding/hex"
+	"encoding/json"
 	"fmt"
+	"io"
 	"net/http"
+	"net/url"
 	"os"
 	"strings"
 	"time"
 
-	"github.com/Mrs4s/go-cqhttp/coolq"
-	"github.com/Mrs4s/go-cqhttp/global/config"
-
 	"github.com/Mrs4s/MiraiGo/utils"
-	"github.com/gin-gonic/gin"
 	"github.com/guonaihong/gout"
 	"github.com/guonaihong/gout/dataflow"
 	log "github.com/sirupsen/logrus"
 	"github.com/tidwall/gjson"
+
+	"github.com/Mrs4s/go-cqhttp/coolq"
+	"github.com/Mrs4s/go-cqhttp/global/config"
 )
 
 type httpServer struct {
-	engine *gin.Engine
-	bot    *coolq.CQBot
-	HTTP   *http.Server
-	api    *apiCaller
+	HTTP        *http.Server
+	api         *apiCaller
+	accessToken string
 }
 
 // HTTPClient 反向HTTP上报客户端
@@ -36,11 +36,117 @@ type HTTPClient struct {
 	secret  string
 	addr    string
 	filter  string
+	apiPort int
 	timeout int32
 }
 
-type httpContext struct {
-	ctx *gin.Context
+type httpCtx struct {
+	json     gjson.Result
+	query    url.Values
+	postForm url.Values
+}
+
+func (h *httpCtx) Get(s string) gjson.Result {
+	j := h.json.Get(s)
+	if j.Exists() {
+		return j
+	}
+	validJSONParam := func(p string) bool {
+		return (strings.HasPrefix(p, "{") || strings.HasPrefix(p, "[")) && gjson.Valid(p)
+	}
+	if h.postForm != nil {
+		if form := h.postForm.Get(s); form != "" {
+			if validJSONParam(form) {
+				return gjson.Result{Type: gjson.JSON, Raw: form}
+			}
+			return gjson.Result{Type: gjson.String, Str: form}
+		}
+	}
+	if h.query != nil {
+		if query := h.query.Get(s); query != "" {
+			if validJSONParam(query) {
+				return gjson.Result{Type: gjson.JSON, Raw: query}
+			}
+			return gjson.Result{Type: gjson.String, Str: query}
+		}
+	}
+	return gjson.Result{}
+}
+
+func (s *httpServer) ServeHTTP(writer http.ResponseWriter, request *http.Request) {
+	var ctx httpCtx
+	contentType := request.Header.Get("Content-Type")
+	switch request.Method {
+	case http.MethodPost:
+		if strings.Contains(contentType, "application/json") {
+			body, err := io.ReadAll(request.Body)
+			if err != nil {
+				log.Warnf("获取请求 %v 的Body时出现错误: %v", request.RequestURI, err)
+				writer.WriteHeader(http.StatusBadRequest)
+				return
+			}
+			if !gjson.ValidBytes(body) {
+				log.Warnf("已拒绝客户端 %v 的请求: 非法Json", request.RemoteAddr)
+				writer.WriteHeader(http.StatusBadRequest)
+				return
+			}
+			ctx.json = gjson.Parse(utils.B2S(body))
+		}
+		if strings.Contains(contentType, "application/x-www-form-urlencoded") {
+			err := request.ParseForm()
+			if err != nil {
+				log.Warnf("已拒绝客户端 %v 的请求: %v", request.RemoteAddr, err)
+				writer.WriteHeader(http.StatusBadRequest)
+			}
+			ctx.postForm = request.PostForm
+		}
+		fallthrough
+	case http.MethodGet:
+		ctx.query = request.URL.Query()
+
+	default:
+		log.Warnf("已拒绝客户端 %v 的请求: 方法错误", request.RemoteAddr)
+		writer.WriteHeader(http.StatusNotFound)
+		return
+	}
+	if status := checkAuth(request, s.accessToken); status != http.StatusOK {
+		writer.WriteHeader(status)
+		return
+	}
+
+	action := strings.TrimPrefix(request.URL.Path, "/")
+	action = strings.TrimSuffix(action, "_async")
+	log.Debugf("HTTPServer接收到API调用: %v", action)
+	ret := s.api.callAPI(action, &ctx)
+
+	writer.Header().Set("Content-Type", "application/json; charset=utf-8")
+	writer.WriteHeader(http.StatusOK)
+	_ = json.NewEncoder(writer).Encode(ret)
+}
+
+func checkAuth(req *http.Request, token string) int {
+	if token == "" { // quick path
+		return http.StatusOK
+	}
+
+	auth := req.Header.Get("Authorization")
+	if auth == "" {
+		auth = req.URL.Query().Get("access_token")
+	} else {
+		authN := strings.SplitN(auth, " ", 2)
+		if len(authN) == 2 {
+			auth = authN[1]
+		}
+	}
+
+	switch auth {
+	case token:
+		return http.StatusOK
+	case "":
+		return http.StatusUnauthorized
+	default:
+		return http.StatusForbidden
+	}
 }
 
 // RunHTTPServerAndClients 启动HTTP服务器与HTTP上报客户端
@@ -49,78 +155,27 @@ func RunHTTPServerAndClients(bot *coolq.CQBot, conf *config.HTTPServer) {
 		return
 	}
 	var (
-		s         = new(httpServer)
-		authToken = conf.AccessToken
-		addr      string
+		s    = new(httpServer)
+		addr string
 	)
+	s.accessToken = conf.AccessToken
 	if conf.Host == "" || conf.Port == 0 {
 		goto client
 	}
 	addr = fmt.Sprintf("%s:%d", conf.Host, conf.Port)
-	gin.SetMode(gin.ReleaseMode)
-	s.engine = gin.New()
-	s.bot = bot
-	s.api = newAPICaller(s.bot)
+	s.api = newAPICaller(bot)
 	if conf.RateLimit.Enabled {
 		s.api.use(rateLimit(conf.RateLimit.Frequency, conf.RateLimit.Bucket))
 	}
-	s.engine.Use(func(c *gin.Context) {
-		if c.Request.Method != "GET" && c.Request.Method != "POST" {
-			log.Warnf("已拒绝客户端 %v 的请求: 方法错误", c.Request.RemoteAddr)
-			c.Status(404)
-			return
-		}
-		if c.Request.Method == "POST" && strings.Contains(c.Request.Header.Get("Content-Type"), "application/json") {
-			d, err := c.GetRawData()
-			if err != nil {
-				log.Warnf("获取请求 %v 的Body时出现错误: %v", c.Request.RequestURI, err)
-				c.Status(400)
-				return
-			}
-			if !gjson.ValidBytes(d) {
-				log.Warnf("已拒绝客户端 %v 的请求: 非法Json", c.Request.RemoteAddr)
-				c.Status(400)
-				return
-			}
-			c.Set("json_body", gjson.ParseBytes(d))
-		}
-		c.Next()
-	})
-
-	if authToken != "" {
-		s.engine.Use(func(c *gin.Context) {
-			auth := c.Request.Header.Get("Authorization")
-			if auth == "" {
-				headAuth := c.Query("access_token")
-				switch {
-				case headAuth == "":
-					c.AbortWithStatus(401)
-					return
-				case headAuth != authToken:
-					c.AbortWithStatus(403)
-					return
-				}
-			} else {
-				auth := strings.SplitN(auth, " ", 2)
-				switch {
-				case len(auth) != 2 || auth[1] == "":
-					c.AbortWithStatus(401)
-					return
-				case auth[1] != authToken:
-					c.AbortWithStatus(403)
-					return
-				}
-			}
-		})
+	if conf.LongPolling.Enabled {
+		s.api.use(longPolling(bot, conf.LongPolling.MaxQueueSize))
 	}
-
-	s.engine.Any("/:action", s.HandleActions)
 
 	go func() {
 		log.Infof("CQ HTTP 服务器已启动: %v", addr)
 		s.HTTP = &http.Server{
 			Addr:    addr,
-			Handler: s.engine,
+			Handler: s,
 		}
 		if err := s.HTTP.ListenAndServe(); err != nil && err != http.ErrServerClosed {
 			log.Error(err)
@@ -133,54 +188,54 @@ func RunHTTPServerAndClients(bot *coolq.CQBot, conf *config.HTTPServer) {
 client:
 	for _, c := range conf.Post {
 		if c.URL != "" {
-			go newHTTPClient().Run(c.URL, c.Secret, conf.Filter, conf.Timeout, bot)
+			go HTTPClient{
+				bot:     bot,
+				secret:  c.Secret,
+				addr:    c.URL,
+				apiPort: conf.Port,
+				filter:  conf.Filter,
+				timeout: conf.Timeout,
+			}.Run()
 		}
 	}
 }
 
-// newHTTPClient 返回反向HTTP客户端
-func newHTTPClient() *HTTPClient {
-	return &HTTPClient{}
-}
-
 // Run 运行反向HTTP服务
-func (c *HTTPClient) Run(addr, secret, filter string, timeout int32, bot *coolq.CQBot) {
-	c.bot = bot
-	c.secret = secret
-	c.addr = addr
-	c.timeout = timeout
-	c.filter = filter
-	addFilter(filter)
+func (c HTTPClient) Run() {
+	addFilter(c.filter)
 	if c.timeout < 5 {
 		c.timeout = 5
 	}
-	bot.OnEventPush(c.onBotPushEvent)
-	log.Infof("HTTP POST上报器已启动: %v", addr)
+	c.bot.OnEventPush(c.onBotPushEvent)
+	log.Infof("HTTP POST上报器已启动: %v", c.addr)
 }
 
-func (c *HTTPClient) onBotPushEvent(m *bytes.Buffer) {
+func (c *HTTPClient) onBotPushEvent(e *coolq.Event) {
 	var res string
 	if c.filter != "" {
 		filter := findFilter(c.filter)
-		if filter != nil && !filter.Eval(gjson.Parse(utils.B2S(m.Bytes()))) {
-			log.Debugf("上报Event %v 到 HTTP 服务器 %v 时被过滤.", c.addr, utils.B2S(m.Bytes()))
+		if filter != nil && !filter.Eval(gjson.Parse(e.JSONString())) {
+			log.Debugf("上报Event %v 到 HTTP 服务器 %s 时被过滤.", c.addr, e.JSONBytes())
 			return
 		}
 	}
 
-	err := gout.POST(c.addr).SetJSON(m.Bytes()).BindBody(&res).SetHeader(func() gout.H {
+	err := gout.POST(c.addr).SetJSON(e.JSONBytes()).BindBody(&res).SetHeader(func() gout.H {
 		h := gout.H{
 			"X-Self-ID":  c.bot.Client.Uin,
 			"User-Agent": "CQHttp/4.15.0",
 		}
 		if c.secret != "" {
 			mac := hmac.New(sha1.New, []byte(c.secret))
-			_, err := mac.Write(m.Bytes())
+			_, err := mac.Write(e.JSONBytes())
 			if err != nil {
 				log.Error(err)
 				return nil
 			}
 			h["X-Signature"] = "sha1=" + hex.EncodeToString(mac.Sum(nil))
+		}
+		if c.apiPort != 0 {
+			h["X-API-Port"] = c.apiPort
 		}
 		return h
 	}()).SetTimeout(time.Second * time.Duration(c.timeout)).F().Retry().Attempt(5).
@@ -193,41 +248,13 @@ func (c *HTTPClient) onBotPushEvent(m *bytes.Buffer) {
 			return nil
 		}).Do()
 	if err != nil {
-		log.Warnf("上报Event数据 %v 到 %v 失败: %v", utils.B2S(m.Bytes()), c.addr, err)
+		log.Warnf("上报Event数据 %s 到 %v 失败: %v", e.JSONBytes(), c.addr, err)
 		return
 	}
-	log.Debugf("上报Event数据 %v 到 %v", utils.B2S(m.Bytes()), c.addr)
+	log.Debugf("上报Event数据 %s 到 %v", e.JSONBytes(), c.addr)
 	if gjson.Valid(res) {
-		c.bot.CQHandleQuickOperation(gjson.Parse(utils.B2S(m.Bytes())), gjson.Parse(res))
+		c.bot.CQHandleQuickOperation(gjson.Parse(e.JSONString()), gjson.Parse(res))
 	}
-}
-
-func (s *httpServer) HandleActions(c *gin.Context) {
-	action := strings.TrimSuffix(c.Param("action"), "_async")
-	log.Debugf("HTTPServer接收到API调用: %v", action)
-	c.JSON(200, s.api.callAPI(action, httpContext{ctx: c}))
-}
-
-func (h httpContext) Get(k string) gjson.Result {
-	c := h.ctx
-	if q := c.Query(k); q != "" {
-		return gjson.Result{Type: gjson.String, Str: q}
-	}
-	if c.Request.Method == "POST" {
-		if h := c.Request.Header.Get("Content-Type"); h != "" {
-			if strings.Contains(h, "application/x-www-form-urlencoded") {
-				if p, ok := c.GetPostForm(k); ok {
-					return gjson.Result{Type: gjson.String, Str: p}
-				}
-			}
-			if strings.Contains(h, "application/json") {
-				if obj, ok := c.Get("json_body"); ok {
-					return obj.(gjson.Result).Get(k)
-				}
-			}
-		}
-	}
-	return gjson.Result{Type: gjson.Null, Str: ""}
 }
 
 func (s *httpServer) ShutDown() {
